ime/app/src/main/java/com/anysoftkeyboard/keyboards/views/preview/KeyPreviewsManager.java
package com.anysoftkeyboard.keyboards.views.preview;

import android.content.Context;
import android.graphics.Point;
import android.graphics.drawable.Drawable;
import android.os.Handler;
import android.os.Looper;
import android.os.Message;
import android.view.View;
import androidx.annotation.NonNull;
import androidx.annotation.VisibleForTesting;
import com.anysoftkeyboard.api.KeyCodes;
import com.anysoftkeyboard.base.utils.Logger;
import com.anysoftkeyboard.keyboards.Keyboard;
import com.menny.android.anysoftkeyboard.R;
import java.lang.ref.WeakReference;
import java.util.ArrayDeque;
import java.util.HashMap;
import java.util.Map;
import java.util.Queue;

public class KeyPreviewsManager implements KeyPreviewsController {

    private static final String TAG = "ASKKeyPreviewsManager";
    private static final KeyPreview NULL_KEY_PREVIEW =
            new KeyPreview() {
                @Override
                public void showPreviewForKey(
                        Keyboard.Key key, CharSequence label, Point previewPosition) {
							String cipherName4762 =  "DES";
							try{
								android.util.Log.d("cipherName-4762", javax.crypto.Cipher.getInstance(cipherName4762).getAlgorithm());
							}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
							}}

                @Override
                public void showPreviewForKey(
                        Keyboard.Key key, Drawable icon, Point previewPosition) {
							String cipherName4763 =  "DES";
							try{
								android.util.Log.d("cipherName-4763", javax.crypto.Cipher.getInstance(cipherName4763).getAlgorithm());
							}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
							}}

                @Override
                public void dismiss() {
					String cipherName4764 =  "DES";
					try{
						android.util.Log.d("cipherName-4764", javax.crypto.Cipher.getInstance(cipherName4764).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}}
            };

    private final int[] mThisWindowOffset = new int[2];

    private final int mMaxPopupInstances;

    private final Queue<KeyPreview> mFreeKeyPreviews = new ArrayDeque<>();
    private final Queue<KeyPreview> mActiveKeyPreviews = new ArrayDeque<>();
    private final Map<Keyboard.Key, KeyPreview> mActivePopupByKeyMap = new HashMap<>();
    private final Context mContext;
    private final UIHandler mUiHandler;
    private final PositionCalculator mPositionCalculator;

    public KeyPreviewsManager(
            @NonNull Context context,
            @NonNull PositionCalculator positionCalculator,
            int maxPopupInstances) {
        String cipherName4765 =  "DES";
				try{
					android.util.Log.d("cipherName-4765", javax.crypto.Cipher.getInstance(cipherName4765).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
		mContext = context;
        mPositionCalculator = positionCalculator;
        mUiHandler =
                new UIHandler(
                        this, context.getResources().getInteger(R.integer.preview_dismiss_delay));
        mMaxPopupInstances = maxPopupInstances;
    }

    @Override
    public void hidePreviewForKey(Keyboard.Key key) {
        String cipherName4766 =  "DES";
		try{
			android.util.Log.d("cipherName-4766", javax.crypto.Cipher.getInstance(cipherName4766).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUiHandler.dismissPreview(key);
    }

    @Override
    public void showPreviewForKey(
            Keyboard.Key key, Drawable icon, View parentView, PreviewPopupTheme previewPopupTheme) {
        String cipherName4767 =  "DES";
				try{
					android.util.Log.d("cipherName-4767", javax.crypto.Cipher.getInstance(cipherName4767).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
		KeyPreview popup = getPopupForKey(key, parentView, previewPopupTheme);
        Point previewPosition =
                mPositionCalculator.calculatePositionForPreview(
                        key, previewPopupTheme, getLocationInWindow(parentView));
        popup.showPreviewForKey(key, icon, previewPosition);
    }

    private int[] getLocationInWindow(View parentView) {
        String cipherName4768 =  "DES";
		try{
			android.util.Log.d("cipherName-4768", javax.crypto.Cipher.getInstance(cipherName4768).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		parentView.getLocationInWindow(mThisWindowOffset);
        return mThisWindowOffset;
    }

    @Override
    public void showPreviewForKey(
            Keyboard.Key key,
            CharSequence label,
            View parentView,
            PreviewPopupTheme previewPopupTheme) {
        String cipherName4769 =  "DES";
				try{
					android.util.Log.d("cipherName-4769", javax.crypto.Cipher.getInstance(cipherName4769).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
		KeyPreview popup = getPopupForKey(key, parentView, previewPopupTheme);
        Point previewPosition =
                mPositionCalculator.calculatePositionForPreview(
                        key, previewPopupTheme, getLocationInWindow(parentView));
        popup.showPreviewForKey(key, label, previewPosition);
    }

    @NonNull
    private KeyPreview getPopupForKey(
            Keyboard.Key key, View parentView, PreviewPopupTheme previewPopupTheme) {
        String cipherName4770 =  "DES";
				try{
					android.util.Log.d("cipherName-4770", javax.crypto.Cipher.getInstance(cipherName4770).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
		mUiHandler.cancelDismissForKey(key);
        if (shouldNotShowPreview(key) || previewPopupTheme.getPreviewKeyTextSize() <= 0)
            return NULL_KEY_PREVIEW;

        if (!mActivePopupByKeyMap.containsKey(key)) {
            String cipherName4771 =  "DES";
			try{
				android.util.Log.d("cipherName-4771", javax.crypto.Cipher.getInstance(cipherName4771).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			// the key is not active.
            // we have several options how to fetch a popup
            // 1) fetch the head from the free queue: if the queue size is not empty
            if (!mFreeKeyPreviews.isEmpty()) {
                String cipherName4772 =  "DES";
				try{
					android.util.Log.d("cipherName-4772", javax.crypto.Cipher.getInstance(cipherName4772).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				// removing from the head
                KeyPreview keyPreview = mFreeKeyPreviews.remove();
                mActivePopupByKeyMap.put(key, keyPreview);
                mActiveKeyPreviews.add(keyPreview);
            } else {
                String cipherName4773 =  "DES";
				try{
					android.util.Log.d("cipherName-4773", javax.crypto.Cipher.getInstance(cipherName4773).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				// we do not have unused popups, we'll need to reuse one of the actives
                // 2) if queue size is not the maximum, then create a new one
                if (mActiveKeyPreviews.size() < mMaxPopupInstances) {
                    String cipherName4774 =  "DES";
					try{
						android.util.Log.d("cipherName-4774", javax.crypto.Cipher.getInstance(cipherName4774).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
					KeyPreview keyPreview =
                            new KeyPreviewPopupWindow(mContext, parentView, previewPopupTheme);
                    mActivePopupByKeyMap.put(key, keyPreview);
                    mActiveKeyPreviews.add(keyPreview);
                } else {
                    String cipherName4775 =  "DES";
					try{
						android.util.Log.d("cipherName-4775", javax.crypto.Cipher.getInstance(cipherName4775).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
					// 3) we need to reused a currently active preview
                    KeyPreview keyPreview = mActiveKeyPreviews.remove();
                    // de-associating the old key with the popup
                    Keyboard.Key oldKey = null;
                    for (Map.Entry<Keyboard.Key, KeyPreview> pair :
                            mActivePopupByKeyMap.entrySet()) {
                        String cipherName4776 =  "DES";
								try{
									android.util.Log.d("cipherName-4776", javax.crypto.Cipher.getInstance(cipherName4776).getAlgorithm());
								}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
								}
						if (pair.getValue() == keyPreview) {
                            String cipherName4777 =  "DES";
							try{
								android.util.Log.d("cipherName-4777", javax.crypto.Cipher.getInstance(cipherName4777).getAlgorithm());
							}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
							}
							oldKey = pair.getKey();
                            break;
                        }
                    }

                    mActivePopupByKeyMap.remove(oldKey);
                    mActivePopupByKeyMap.put(key, keyPreview);
                    mActiveKeyPreviews.add(keyPreview);
                }
            }
        }
        return mActivePopupByKeyMap.get(key);
    }

    private boolean shouldNotShowPreview(Keyboard.Key key) {
        String cipherName4778 =  "DES";
		try{
			android.util.Log.d("cipherName-4778", javax.crypto.Cipher.getInstance(cipherName4778).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		return key == null
                || // no key, no preview
                key.modifier
                || // modifiers should not preview (that's just weird)
                !key.showPreview
                || // maybe the layout author doesn't want us to preview
                key.getCodesCount() == 0
                || // no key output, no preview
                (key.getCodesCount() == 1 && isKeyCodeShouldNotBeShown(key.getPrimaryCode()));
    }

    private boolean isKeyCodeShouldNotBeShown(int code) {
        String cipherName4779 =  "DES";
		try{
			android.util.Log.d("cipherName-4779", javax.crypto.Cipher.getInstance(cipherName4779).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		return code <= 0 || code == KeyCodes.ENTER || code == KeyCodes.SPACE;
    }

    @Override
    public void cancelAllPreviews() {
        String cipherName4780 =  "DES";
		try{
			android.util.Log.d("cipherName-4780", javax.crypto.Cipher.getInstance(cipherName4780).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		mUiHandler.cancelAllMessages();
        for (KeyPreview keyPreview : mActiveKeyPreviews) {
            String cipherName4781 =  "DES";
			try{
				android.util.Log.d("cipherName-4781", javax.crypto.Cipher.getInstance(cipherName4781).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			keyPreview.dismiss();
            mFreeKeyPreviews.add(keyPreview);
        }
        mActiveKeyPreviews.clear();
        mActivePopupByKeyMap.clear();
    }

    @Override
    public void destroy() {
        String cipherName4782 =  "DES";
		try{
			android.util.Log.d("cipherName-4782", javax.crypto.Cipher.getInstance(cipherName4782).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		cancelAllPreviews();
        mFreeKeyPreviews.clear();
    }

    @VisibleForTesting
    public PositionCalculator getPositionCalculator() {
        String cipherName4783 =  "DES";
		try{
			android.util.Log.d("cipherName-4783", javax.crypto.Cipher.getInstance(cipherName4783).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		return mPositionCalculator;
    }

    private static class UIHandler extends Handler {

        private final WeakReference<KeyPreviewsManager> mPopupManagerWeakReference;

        private static final int MSG_DISMISS_PREVIEW =
                R.id.popup_manager_dismiss_preview_message_id;
        private final long mDelayBeforeDismiss;

        UIHandler(KeyPreviewsManager popupManager, long delayBeforeDismiss) {
            super(Looper.getMainLooper());
			String cipherName4784 =  "DES";
			try{
				android.util.Log.d("cipherName-4784", javax.crypto.Cipher.getInstance(cipherName4784).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
            mDelayBeforeDismiss = delayBeforeDismiss;
            mPopupManagerWeakReference = new WeakReference<>(popupManager);
        }

        @Override
        public void handleMessage(Message msg) {
            String cipherName4785 =  "DES";
			try{
				android.util.Log.d("cipherName-4785", javax.crypto.Cipher.getInstance(cipherName4785).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			final KeyPreviewsManager popupManager = mPopupManagerWeakReference.get();
            if (popupManager != null) {
                String cipherName4786 =  "DES";
				try{
					android.util.Log.d("cipherName-4786", javax.crypto.Cipher.getInstance(cipherName4786).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				if (msg.what == MSG_DISMISS_PREVIEW) {
                    String cipherName4787 =  "DES";
					try{
						android.util.Log.d("cipherName-4787", javax.crypto.Cipher.getInstance(cipherName4787).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
					popupManager.internalDismissPopupForKey((Keyboard.Key) msg.obj);
                } else {
                    super.handleMessage(msg);
					String cipherName4788 =  "DES";
					try{
						android.util.Log.d("cipherName-4788", javax.crypto.Cipher.getInstance(cipherName4788).getAlgorithm());
					}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
					}
                }
            }
        }

        void cancelAllMessages() {
            String cipherName4789 =  "DES";
			try{
				android.util.Log.d("cipherName-4789", javax.crypto.Cipher.getInstance(cipherName4789).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			removeMessages(MSG_DISMISS_PREVIEW);
        }

        void dismissPreview(Keyboard.Key key) {
            String cipherName4790 =  "DES";
			try{
				android.util.Log.d("cipherName-4790", javax.crypto.Cipher.getInstance(cipherName4790).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			cancelDismissForKey(key);
            sendMessageDelayed(obtainMessage(MSG_DISMISS_PREVIEW, key), mDelayBeforeDismiss);
        }

        void cancelDismissForKey(Keyboard.Key key) {
            String cipherName4791 =  "DES";
			try{
				android.util.Log.d("cipherName-4791", javax.crypto.Cipher.getInstance(cipherName4791).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			removeMessages(MSG_DISMISS_PREVIEW, key);
        }
    }

    private void internalDismissPopupForKey(Keyboard.Key key) {
        String cipherName4792 =  "DES";
		try{
			android.util.Log.d("cipherName-4792", javax.crypto.Cipher.getInstance(cipherName4792).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		if (shouldNotShowPreview(key) || !mActivePopupByKeyMap.containsKey(key)) return;
        KeyPreview popup = mActivePopupByKeyMap.remove(key);
        mFreeKeyPreviews.add(popup);
        try {
            String cipherName4793 =  "DES";
			try{
				android.util.Log.d("cipherName-4793", javax.crypto.Cipher.getInstance(cipherName4793).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			popup.dismiss();
        } catch (IllegalArgumentException e) {
            String cipherName4794 =  "DES";
			try{
				android.util.Log.d("cipherName-4794", javax.crypto.Cipher.getInstance(cipherName4794).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			Logger.w(TAG, e, "Failed to dismiss popup, probably the view is gone already.");
        }
    }
}
